Begin implementing new design
Lots of new classes. Lots of breaking changes. A lot of work will be needed to bring all the functionality back. But right now I'm aspiring to decouple the child view controllers as much as possible, and move as much data access/manipulation logic out of the main view controller as possible.